[IMP] tools: change bundle scripts to ES6
Changed the bundle scripts from commonJS modules (cjs) to ES6 modules.
This allow us to use the more modern ES6 syntax and features.

Notable changes :
- replace `require` by `import`
- replace `module.exports` by `export function ...`
- cannot import a json file directly (experimental feature), read it
and parse it with `fs` and `JSON.parse` instead.
- __dirname is not available anymore, use `import.meta.url` instead

This also required to setup babel to transpile the ES6 modules to
commonJS modules for jest.
